server/requirements.txt
asgiref==3.10.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
coverage==7.11.0
cryptography==46.0.3
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.12.1
django-extensions==4.1
djangorestframework==3.16.1
google-auth==2.41.1
idna==3.11
iniconfig==2.3.0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
